pyproject.toml
[tool.poetry]
name = "fastapi-django"
version = "0.1.0"
description = "Test django orm with eaas"
authors = ["Globex Corporation"]
license = "Apache Software License 2.0"
packages = [
    { include = "connect_ext" }
]
readme = "./README.md"

[tool.poetry.plugins."connect.eaas.ext"]
"webapp" = "connect_ext.webapp:FastapiDjangoWebApplication"
"eventsapp" = "connect_ext.events:E2EEventsApplication"
"djsettings" = "connect_ext:get_django_settings_module"

[tool.poetry.dependencies]
python = ">=3.8,<4"
connect-eaas-core = ">=28.12,<29"
django = "^4.2.3"
psycopg2-binary = "^2.9.6"
django-rql = "^4.4.0"

[tool.poetry.dev-dependencies]
pytest = ">=6.1.2,<8"
pytest-cov = ">=2.10.1,<5"
pytest-mock = "^3.3.1"
mock = { version = "^4.0.3", markers = "python_version < '3.8'" }
coverage = {extras = ["toml"], version = ">=5.3,<7"}
flake8 = ">=3.8,<6"
flake8-bugbear = ">=20,<23"
flake8-cognitive-complexity = "^0.1"
flake8-commas = "~2.0"
flake8-future-import = "~0.4"
flake8-import-order = "~0.18"
flake8-broken-line = ">=0.3,<0.7"
flake8-comprehensions = "^3.3.1"
flake8-debugger = "^4.0.0"
flake8-eradicate = "^1.0.0"
flake8-string-format = "^0.3.0"
pytest-asyncio = "^0.15.1"

[tool.poetry.group.dev.dependencies]
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = "tests"
addopts = "--cov=connect_ext --cov-report=term-missing --cov-report=html --cov-report=xml"
DJANGO_SETTINS_MODULE = "connect_ext.django.settings"
django_find_project = false
pythonpath = [".", "connect_ext"]

[tool.coverage.run]
relative_files = true
branch = true

[tool.coverage.report]
omit = [
]
